feat: deb and rpm packages (#46)

+++ res/scripts/rpm/preremove.sh
@@ -4,7 +4,7 @@ set -e
 set -x
 
 echo "Nord Sense Linux preremove.sh script"
-executable=(readlink /proc/1/exe)
+executable=$(readlink /proc/1/exe)
 bin=${executable##*/}
 if [ "$bin" = "systemd" ]; then
     if ! which systemctl 2>&1 1>/dev/null; then
@@ -16,10 +16,12 @@ if [ "$bin" = "systemd" ]; then
         systemctl stop nsense
     fi
 
-    rm -f /etc/systemd/system/npsense*.service
+    if [ $(find /etc/systemd/system -iname "nsense" | wc -l) -gt 0 ]; then
+        systemctl disable nsense
+        systemctl disable nsense-sleep
+        rm -f /etc/systemd/system/npsense*.service
+    fi
     systemctl daemon-reload
-    systemctl disable nsense
-    systemctl disable nsense-sleep
 else
     echo "ERROR: your init system seems isn't supported" 1>&2
     exit 1
